cambio int a long y comentarios

+++ src/main/java/api/dao/IHotelDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import api.dto.Hotel;
 
-public interface IHotelDAO extends JpaRepository<Hotel, Integer> {
+public interface IHotelDAO extends JpaRepository<Hotel, Long> {
 
 }
