+++ src/main/java/api/service/IHotelService.java
@@ -5,14 +5,15 @@ import java.util.List;
 import api.dto.Hotel;
 
 public interface IHotelService {
-
+		
+	//Metodos del CRUD
 		public List<Hotel> listarHotel();
 		
 		public Hotel guardarHotel(Hotel hoteles);
 		
-		public Hotel HotelXID(Integer id);
+		public Hotel HotelXID(Long id);
 		
 		public Hotel actualizarHotel(Hotel hoteles);
 	
-		public void eliminarHotel(Integer id);
+		public void eliminarHotel(Long id);
 }
